# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminContactMessageResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "NEW", "READ", "PROCESSED", "ARCHIVED" ],
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "email", "id", "message", "name", "status", "subject", "updatedAt" ],
        "type" : "object"
      },
      "AdminContactMessageSummaryResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "NEW", "READ", "PROCESSED", "ARCHIVED" ],
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "email", "id", "name", "status", "subject" ],
        "type" : "object"
      },
...
      "AdminPublicationResponse" : {
        "properties" : {
          "categoryId" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "contentMarkdown" : {
            "type" : "string"
          },
          "coverMediaId" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "featured" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "seoDescription" : {
            "type" : [ "string", "null" ]
          },
          "seoTitle" : {
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : "string"
          },
          "slugLocked" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [ "DRAFT", "IN_REVIEW", "SCHEDULED", "PUBLISHED", "ARCHIVED" ],
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "tagIds" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "ARTICLE", "NEWS" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "categoryId", "contentMarkdown", "coverMediaId", "createdAt", "featured", "id", "publishedAt", "seoDescription", "seoTitle", "slug", "slugLocked", "status", "summary", "tagIds", "title", "type", "updatedAt" ],
        "type" : "object"
      },
      "AdminPublicationSummaryResponse" : {
        "properties" : {
          "featured" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "IN_REVIEW", "SCHEDULED", "PUBLISHED", "ARCHIVED" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "ARTICLE", "NEWS" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "featured", "id", "publishedAt", "slug", "status", "title", "type", "updatedAt" ],
        "type" : "object"
      },
...
      "ChangeContactMessageStatusRequest" : {
        "properties" : {
          "status" : {
            "enum" : [ "NEW", "READ", "PROCESSED", "ARCHIVED" ],
            "type" : "string"
          }
        },
        "required" : [ "status" ],
        "type" : "object"
      },
      "ChangePublicationStatusRequest" : {
        "properties" : {
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "IN_REVIEW", "SCHEDULED", "PUBLISHED", "ARCHIVED" ],
            "type" : "string"
          }
        },
        "required" : [ "status" ],
        "type" : "object"
      },
...
      "CreatePublicationRequest" : {
        "properties" : {
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "contentMarkdown" : {
            "maxLength" : 100000,
            "minLength" : 0,
            "type" : "string"
          },
          "coverMediaId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "featured" : {
            "type" : "boolean"
          },
          "seoDescription" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "seoTitle" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "slug" : {
            "maxLength" : 160,
            "minLength" : 0,
            "pattern" : "[a-z0-9]+(-[a-z0-9]+)*",
            "type" : "string"
          },
          "summary" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "tagIds" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "title" : {
            "maxLength" : 160,
            "minLength" : 0,
            "pattern" : "(?s).*[\\p{L}\\p{N}].*",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "ARTICLE", "NEWS" ],
            "type" : "string"
          }
        },
        "required" : [ "contentMarkdown", "summary", "title", "type" ],
        "type" : "object"
      },
...
      "PageResponseAdminContactMessageSummaryResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/AdminContactMessageSummaryResponse"
            },
            "type" : "array"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "first", "last", "page", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
...
      "PageResponseAdminPublicationSummaryResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/AdminPublicationSummaryResponse"
            },
            "type" : "array"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "first", "last", "page", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "PublicImage" : {
        "properties" : {
          "altText" : {
            "type" : [ "string", "null" ]
          },
          "height" : {
            "format" : "int32",
            "type" : "integer"
          },
          "url" : {
            "type" : "string"
          },
          "width" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "altText", "height", "url", "width" ],
        "type" : "object"
      },
      "PublicationResponse" : {
        "properties" : {
          "category" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/TaxonomyTermResponse"
            }, {
              "type" : "null"
            } ]
          },
          "contentMarkdown" : {
            "type" : "string"
          },
          "cover" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/PublicImage"
            }, {
              "type" : "null"
            } ]
          },
          "featured" : {
            "type" : "boolean"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "readingTimeMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "seoDescription" : {
            "type" : [ "string", "null" ]
          },
          "seoTitle" : {
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "$ref" : "#/components/schemas/TaxonomyTermResponse"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "ARTICLE", "NEWS" ],
            "type" : "string"
          }
        },
        "required" : [ "category", "contentMarkdown", "cover", "featured", "publishedAt", "readingTimeMinutes", "seoDescription", "seoTitle", "slug", "summary", "tags", "title", "type" ],
        "type" : "object"
      },
...
      "TaxonomyTermResponse" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        },
        "required" : [ "name", "slug" ],
        "type" : "object"
      },
...
      "UpdatePublicationRequest" : {
        "properties" : {
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "contentMarkdown" : {
            "maxLength" : 100000,
            "minLength" : 0,
            "type" : "string"
          },
          "coverMediaId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "featured" : {
            "type" : "boolean"
          },
          "seoDescription" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "seoTitle" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "slug" : {
            "maxLength" : 160,
            "minLength" : 0,
            "pattern" : "[a-z0-9]+(-[a-z0-9]+)*",
            "type" : "string"
          },
          "summary" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "tagIds" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "title" : {
            "maxLength" : 160,
            "minLength" : 0,
            "pattern" : "(?s).*[\\p{L}\\p{N}].*",
            "type" : "string"
          }
        },
        "required" : [ "contentMarkdown", "summary", "title" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/contact-messages" : {
      "get" : {
        "operationId" : "list_7",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "NEW", "READ", "PROCESSED", "ARCHIVED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseAdminContactMessageSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-contact-message-controller" ]
      }
    },
    "/admin/contact-messages/{id}" : {
      "delete" : {
        "operationId" : "delete_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "admin-contact-message-controller" ]
      },
      "get" : {
        "operationId" : "get_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminContactMessageResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-contact-message-controller" ]
      }
    },
    "/admin/contact-messages/{id}/status" : {
      "post" : {
        "operationId" : "changeStatus_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeContactMessageStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminContactMessageResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-contact-message-controller" ]
      }
    },
...
    "/admin/publications" : {
      "get" : {
        "operationId" : "list_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "type",
          "required" : false,
          "schema" : {
            "enum" : [ "ARTICLE", "NEWS" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "DRAFT", "IN_REVIEW", "SCHEDULED", "PUBLISHED", "ARCHIVED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseAdminPublicationSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-publication-controller" ]
      },
      "post" : {
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePublicationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPublicationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-publication-controller" ]
      }
    },
    "/admin/publications/{id}" : {
      "get" : {
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPublicationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-publication-controller" ]
      },
      "put" : {
        "operationId" : "update_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdatePublicationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPublicationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-publication-controller" ]
      }
    },
    "/admin/publications/{id}/status" : {
      "post" : {
        "operationId" : "changeStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePublicationStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPublicationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-publication-controller" ]
      }
    },
...
    "/public/publications/{slug}" : {
      "get" : {
        "operationId" : "getPublication",
        "parameters" : [ {
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "public-publication-controller" ]
      }
    },
...
  }
}